FIXED:  Attachments. Working on my small screen, I first did not see the “Upload Your Data” section, and thought that “attachments” was for importing the data.  So “attachments” should explain what type of attachment is intended.  For instance, should annotations go here?
o   Attachments – “Add Files”.  When I did use this function to attach a file, it uploaded a file but then left me stuck on that page. There seems to be no way to navigate beyond that page except with the “back” button on the browser. Should it not have navigation functionality>
, refs #302