Fixes for
Return to Submit New Data page: the user is returned to the middle of the page, without a clear statement of where one is. Perhaps a statement can be added that one has returned to the  Submit New Data page.

o   The “Data Preview” on the Submit New Data page does not now present any information.

o   The “Finish your Submission” button is hard to find – it is too far down on the page and needs to be marked more clearly.

o   The “Connect to the Puzzle” section works very nicely in showing the relationships (including duplicate files in my case). Would it make sense to place the “Finish your Submission” button ahead of Step 3, and then create an “Update your Submission” button for those who have added new relationships in Stage 3?

Typo in second paragraph:

… “crowdsource” the large-scale historical data …

refs #302